+++ frontend/src/components/EquipmentDetails.js
@@ -1,11 +1,20 @@
 import { useEquipmentContext } from '../hooks/useEquipmentContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 const EquipmentDetails = ({ equipment }) => {
     const { dispatch } = useEquipmentContext()
+    const { user } = useAuthContext()
  
     const handleClick = async() => {
+        if (!user) {
+            return
+        }
+
         const response = await fetch('/api/equipment/' + equipment._id, {
-            method: "DELETE"
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
 
@@ -16,7 +25,7 @@ const EquipmentDetails = ({ equipment }) => {
 
     return (
         <div className="form-details">
-            <button className="delete-button" onClick={handleClick}> Usuń sprzęt </button>
+            {user.role === 1 && <button className="delete-button" onClick={handleClick}> Usuń sprzęt </button>}
             <p><strong>Nazwa: </strong>{equipment.nazwa}</p> 
             <p><strong>Kategoria: </strong>{equipment.kategoria}</p> 
             <p><strong>Liczba sprzętu: </strong>{equipment.liczba_sprzetu}</p>
